dash_analytics/requirements.txt
django==4.2.20
djangorestframework>=3.14.0
django-tailwind==3.6.0
whitenoise==6.5.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
prophet>=1.1.4
mongoengine==0.27.0
django-browser-reload==1.11.0
django-cors-headers==4.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
python-decouple==3.8
requests>=2.31.0
djangorestframework-simplejwt==5.3.0
pymongo==4.5.0
bcrypt==4.0.1
setuptools>=79.0.1
blinker==1.6.0